# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lex Intelligentia - Legal AI Operating System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LegalCase {
  id: number;
  titulo: string;
  fatos: string;
  partes: string;
  problema: string;
  objetivo: string;
  /** @nullable */
  documentos?: string | null;
  status: string;
  /** @nullable */
  tipo?: string | null;
  /** @nullable */
  numero_processo?: string | null;
  created_at: string;
  updated_at: string;
}

export interface CreateLegalCaseBody {
  titulo: string;
  fatos: string;
  partes: string;
  problema: string;
  objetivo: string;
  /** @nullable */
  documentos?: string | null;
  /** @nullable */
  tipo?: string | null;
  /** @nullable */
  numero_processo?: string | null;
}

export interface UpdateLegalCaseBody {
  titulo?: string;
  fatos?: string;
  partes?: string;
  problema?: string;
  objetivo?: string;
  /** @nullable */
  documentos?: string | null;
  status?: string;
  /** @nullable */
  tipo?: string | null;
  /** @nullable */
  numero_processo?: string | null;
}

export interface AnalyzeCaseBody {
  skill_id: string;
  skill_name: string;
  /** @nullable */
  additional_context?: string | null;
}

export interface RunWorkflowBody {
  /** sentenca_civel | saneamento | despacho | conciliacao */
  workflow_type: string;
  /** @nullable */
  additional_context?: string | null;
}

export interface AnalysisResult {
  id: number;
  case_id: number;
  /** @nullable */
  skill_id?: string | null;
  /** @nullable */
  skill_name?: string | null;
  /** @nullable */
  workflow_type?: string | null;
  analise: string;
  fundamentacao_juridica: string;
  jurisprudencia_nivel1: string;
  jurisprudencia_nivel2: string;
  jurisprudencia_nivel3: string;
  decisao_recomendacao: string;
  status: string;
  created_at: string;
}

export interface CaseTypeCount {
  tipo: string;
  count: number;
}

export interface RecentActivity {
  id: number;
  description: string;
  timestamp: string;
  type: string;
}

export interface DashboardStats {
  total_cases: number;
  active_cases: number;
  completed_analyses: number;
  pending_cases: number;
  cases_by_type: CaseTypeCount[];
  recent_activity: RecentActivity[];
  workflows_executed: number;
}

export interface LegalSkill {
  id: string;
  name: string;
  description: string;
  category: string;
  icon: string;
}

export type ListCasesParams = {
  status?: string;
  limit?: number;
  offset?: number;
};

export type ListResultsParams = {
  caseId?: number;
};