# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lex Intelligentia - Legal AI Operating System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all cases
 */
export const ListCasesQueryParams = zod.object({
  status: zod.coerce.string().optional(),
  limit: zod.coerce.number().optional(),
  offset: zod.coerce.number().optional(),
});

export const ListCasesResponseItem = zod.object({
  id: zod.number(),
  titulo: zod.string(),
  fatos: zod.string(),
  partes: zod.string(),
  problema: zod.string(),
  objetivo: zod.string(),
  documentos: zod.string().nullish(),
  status: zod.string(),
  tipo: zod.string().nullish(),
  numero_processo: zod.string().nullish(),
  created_at: zod.string(),
  updated_at: zod.string(),
});
export const ListCasesResponse = zod.array(ListCasesResponseItem);

/**
 * @summary Create a new legal case
 */
export const CreateCaseBody = zod.object({
  titulo: zod.string(),
  fatos: zod.string(),
  partes: zod.string(),
  problema: zod.string(),
  objetivo: zod.string(),
  documentos: zod.string().nullish(),
  tipo: zod.string().nullish(),
  numero_processo: zod.string().nullish(),
});

/**
 * @summary Get a legal case by ID
 */
export const GetCaseParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCaseResponse = zod.object({
  id: zod.number(),
  titulo: zod.string(),
  fatos: zod.string(),
  partes: zod.string(),
  problema: zod.string(),
  objetivo: zod.string(),
  documentos: zod.string().nullish(),
  status: zod.string(),
  tipo: zod.string().nullish(),
  numero_processo: zod.string().nullish(),
  created_at: zod.string(),
  updated_at: zod.string(),
});

/**
 * @summary Update a legal case
 */
export const UpdateCaseParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateCaseBody = zod.object({
  titulo: zod.string().optional(),
  fatos: zod.string().optional(),
  partes: zod.string().optional(),
  problema: zod.string().optional(),
  objetivo: zod.string().optional(),
  documentos: zod.string().nullish(),
  status: zod.string().optional(),
  tipo: zod.string().nullish(),
  numero_processo: zod.string().nullish(),
});

export const UpdateCaseResponse = zod.object({
  id: zod.number(),
  titulo: zod.string(),
  fatos: zod.string(),
  partes: zod.string(),
  problema: zod.string(),
  objetivo: zod.string(),
  documentos: zod.string().nullish(),
  status: zod.string(),
  tipo: zod.string().nullish(),
  numero_processo: zod.string().nullish(),
  created_at: zod.string(),
  updated_at: zod.string(),
});

/**
 * @summary Delete a legal case
 */
export const DeleteCaseParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Run AI analysis on a case with a selected skill
 */
export const AnalyzeCaseParams = zod.object({
  id: zod.coerce.number(),
});

export const AnalyzeCaseBody = zod.object({
  skill_id: zod.string(),
  skill_name: zod.string(),
  additional_context: zod.string().nullish(),
});

export const AnalyzeCaseResponse = zod.object({
  id: zod.number(),
  case_id: zod.number(),
  skill_id: zod.string().nullish(),
  skill_name: zod.string().nullish(),
  workflow_type: zod.string().nullish(),
  analise: zod.string(),
  fundamentacao_juridica: zod.string(),
  jurisprudencia_nivel1: zod.string(),
  jurisprudencia_nivel2: zod.string(),
  jurisprudencia_nivel3: zod.string(),
  decisao_recomendacao: zod.string(),
  status: zod.string(),
  created_at: zod.string(),
});

/**
 * @summary Execute a predefined legal workflow on a case
 */
export const RunWorkflowParams = zod.object({
  id: zod.coerce.number(),
});

export const RunWorkflowBody = zod.object({
  workflow_type: zod
    .string()
    .describe("sentenca_civel | saneamento | despacho | conciliacao"),
  additional_context: zod.string().nullish(),
});

export const RunWorkflowResponse = zod.object({
  id: zod.number(),
  case_id: zod.number(),
  skill_id: zod.string().nullish(),
  skill_name: zod.string().nullish(),
  workflow_type: zod.string().nullish(),
  analise: zod.string(),
  fundamentacao_juridica: zod.string(),
  jurisprudencia_nivel1: zod.string(),
  jurisprudencia_nivel2: zod.string(),
  jurisprudencia_nivel3: zod.string(),
  decisao_recomendacao: zod.string(),
  status: zod.string(),
  created_at: zod.string(),
});

/**
 * @summary List all analysis results
 */
export const ListResultsQueryParams = zod.object({
  caseId: zod.coerce.number().optional(),
});

export const ListResultsResponseItem = zod.object({
  id: zod.number(),
  case_id: zod.number(),
  skill_id: zod.string().nullish(),
  skill_name: zod.string().nullish(),
  workflow_type: zod.string().nullish(),
  analise: zod.string(),
  fundamentacao_juridica: zod.string(),
  jurisprudencia_nivel1: zod.string(),
  jurisprudencia_nivel2: zod.string(),
  jurisprudencia_nivel3: zod.string(),
  decisao_recomendacao: zod.string(),
  status: zod.string(),
  created_at: zod.string(),
});
export const ListResultsResponse = zod.array(ListResultsResponseItem);

/**
 * @summary Get an analysis result by ID
 */
export const GetResultParams = zod.object({
  id: zod.coerce.number(),
});

export const GetResultResponse = zod.object({
  id: zod.number(),
  case_id: zod.number(),
  skill_id: zod.string().nullish(),
  skill_name: zod.string().nullish(),
  workflow_type: zod.string().nullish(),
  analise: zod.string(),
  fundamentacao_juridica: zod.string(),
  jurisprudencia_nivel1: zod.string(),
  jurisprudencia_nivel2: zod.string(),
  jurisprudencia_nivel3: zod.string(),
  decisao_recomendacao: zod.string(),
  status: zod.string(),
  created_at: zod.string(),
});

/**
 * @summary Delete an analysis result
 */
export const DeleteResultParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  total_cases: zod.number(),
  active_cases: zod.number(),
  completed_analyses: zod.number(),
  pending_cases: zod.number(),
  cases_by_type: zod.array(
    zod.object({
      tipo: zod.string(),
      count: zod.number(),
    }),
  ),
  recent_activity: zod.array(
    zod.object({
      id: zod.number(),
      description: zod.string(),
      timestamp: zod.string(),
      type: zod.string(),
    }),
  ),
  workflows_executed: zod.number(),
});

/**
 * @summary List all available legal skills
 */
export const ListSkillsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  icon: zod.string(),
});
export const ListSkillsResponse = zod.array(ListSkillsResponseItem);